requirements.txt
django
djangorestframework
celery
redis
requests